Fixed buggy sorting (significant changes)
Testing the sort on longer arrays revealed a buggy sorting order, this
has now been fixed, resulting in several significant updates in the
process (though not all updates are strictly related to the fix, they
are included in this commit anyway).
Note the sort is currently still performing too many comparisons.